backend/requirements.txt
# Core
fastapi==0.109.2
uvicorn[standard]==0.27.1
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1

# Database
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1
redis==5.0.1

# Validation
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.0.post1

# External APIs
httpx==0.26.0
aiohttp==3.9.3

# Utils
qrcode[pil]==7.4.2
pillow==10.2.0
python-dateutil==2.8.2

# Development
pytest==7.4.4
pytest-asyncio==0.23.4
pytest-cov==4.1.0
black==24.1.1
flake8==7.0.0
mypy==1.8.0

# Additional dependencies for stability
bcrypt==4.1.2
cryptography==42.0.2
